+++ asdf/util.py
@@ -267,6 +267,24 @@ def resolve_name(name):
     return ret
 
 
+def get_class_name(obj, instance=True):
+    """
+    Given a class or instance of a class, returns a string representing the
+    fully specified path of the class.
+
+    Parameters
+    ----------
+
+    obj : object
+        An instance of any object
+    instance: bool
+        Indicates whether given object is an instance of the class to be named
+    """
+
+    typ = type(obj) if instance else obj
+    return "{}.{}".format(typ.__module__, typ.__name__)
+
+
 def minversion(module, version, inclusive=True, version_path='__version__'):
     """
     Returns `True` if the specified Python module satisfies a minimum version